+++ CHANGELOG.md
@@ -7,6 +7,7 @@ All notable changes to `@homebridge/hap-nodejs` will be documented in this file.
 ### Changes
 
 - feat: log a debug line when the accessory configuration number increments
+- fix: hash a canonical form of the configuration so array reordering cannot increment c#
 
 ### Homebridge Dependencies